requirements.txt
langchain
langgraph
langchain_community
langchain_core
langchain_groq
langchain_openai
faiss-cpu
streamlit
pandas 
openai
plotly
bcrypt
duckduckgo-search